requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
openai>=1.0.0
easyocr
pytesseract
numpy
opencv-python-headless
